refactor react to post
