+++ src/js/emoji.js
@@ -1,5 +1,8 @@
 import createElementFactory from "./createElementFactory.js";
 export default function createSmileyPicker(parentElement = document.querySelector("main")) {
+    if (document.querySelector("#modal")) {
+        return;
+    }
     const emojiList = [
         { emoji: "😀", altText: "Grinning Face" },
         { emoji: "😃", altText: "Grinning Face with Big Eyes" },
