+++ src/js/singlePost.js
@@ -14,6 +14,7 @@ import renderPosts from "./renderPost.js";
 import commentButton from "./commentOnClick.js";
 import deletePost from "./deleteOnClick.js";
 import updatePost from "./updateOnClick.js";
+import reactToPostTwo from "./reactToPost.js";
 const params = new URLSearchParams(window.location.search);
 const id = params.get("id");
 const endpoint = endpointObject("Jarle");
@@ -25,4 +26,5 @@ const getId = optionFactory("GET", {}, endpoint);
     commentButton();
     deletePost();
     updatePost();
+    reactToPostTwo();
 }))();
